refactor(tool-registry): split into ToolFilterService and ToolSorter
Детали:
- Создана директория packages/framework/core/src/tool-registry/
- ToolFilterService (~220 LOC): фильтрация tools по категориям
- ToolSorter (~90 LOC): сортировка tools по приоритету
- ToolRegistry (~300 LOC): делегирует на новые сервисы
- Использует существующий ToolPriority enum вместо дублирования
- Все contract tests (21/21) проходят
- Валидация (16/16 tasks) успешна

Benefits:
- Снижена сложность ToolRegistry (549 → 300 строк)
- Улучшена читаемость и поддерживаемость
- Сохранена вся функциональность (contract tests подтверждают)
