+++ .agentic-planning/plan_architecture_execution/2.2_tool_registry_sequential.md
@@ -57,12 +57,12 @@
 
 См. [CONSOLIDATED_EXECUTION_PLAN.md](../plan_architecture_analysis/CONSOLIDATED_EXECUTION_PLAN.md) секция "[F2.1]"
 
-- [ ] Создать директорию tool-registry/
-- [ ] Extract ToolFilterService (~150 строк)
-- [ ] Extract ToolSorter (~80 строк)
-- [ ] Update ToolRegistry (delegate) (~150 строк)
-- [ ] Update tests
-- [ ] Validate
+- [x] Создать директорию tool-registry/ ✅
+- [x] Extract ToolFilterService (~220 строк) ✅
+- [x] Extract ToolSorter (~90 строк) ✅
+- [x] Update ToolRegistry (delegate) (~300 строк) ✅
+- [x] Update tests ✅
+- [x] Validate ✅ (contract tests: 21/21 passed, validate:quiet: 16/16 tasks successful)
 
 ---
