+++ packages/framework/core/src/tool-registry/index.ts
@@ -1,0 +1,9 @@
+/**
+ * ToolRegistry - публичный API
+ */
+
+export { ToolRegistry } from './tool-registry.js';
+export { ToolFilterService } from './tool-filter.service.js';
+export { ToolSorter } from './tool-sorter.js';
+export type { ToolConstructor } from './types.js';
+export { PRIORITY_ORDER } from './types.js';
