+++ packages/framework/core/src/tool-registry/tool-registry.ts
@@ -5,6 +5,7 @@
  * - Регистрация инструментов
  * - Получение списка определений
  * - Маршрутизация вызовов к нужному инструменту
+ * - Делегирование фильтрации и сортировки специализированным сервисам
  *
  * АВТОМАТИЧЕСКАЯ РЕГИСТРАЦИЯ (Open/Closed Principle):
  * - Tools автоматически извлекаются из DI контейнера
@@ -19,15 +20,10 @@ import type {
   ToolResult,
   ParsedCategoryFilter,
 } from '@mcp-framework/infrastructure';
-import type { BaseTool, ToolDefinition } from './tools/base/index.js';
-
-/**
- * Конструктор класса Tool для DI
- */
-export interface ToolConstructor {
-  new (...args: any[]): BaseTool<any>; // eslint-disable-line @typescript-eslint/no-explicit-any
-  name: string;
-}
+import type { BaseTool, ToolDefinition } from '../tools/base/index.js';
+import type { ToolConstructor } from './types.js';
+import { ToolFilterService } from './tool-filter.service.js';
+import { ToolSorter } from './tool-sorter.js';
 
 /**
  * Реестр инструментов
@@ -39,6 +35,8 @@ export class ToolRegistry {
   private readonly container: Container;
   private readonly logger: Logger;
   private readonly toolClasses: readonly ToolConstructor[];
+  private readonly filterService: ToolFilterService;
+  private readonly sorter: ToolSorter;
 
   /**
    * @param container - DI контейнер с зарегистрированными tools
@@ -49,6 +47,8 @@ export class ToolRegistry {
     this.container = container;
     this.logger = logger;
     this.toolClasses = toolClasses;
+    this.filterService = new ToolFilterService(logger);
+    this.sorter = new ToolSorter(logger);
     // Не инициализируем tools сразу — делаем это lazy
   }
 
@@ -103,42 +103,6 @@ export class ToolRegistry {
   }
 
   /**
-   * Сортировка инструментов по приоритету
-   *
-   * Порядок: critical → high → normal → low → алфавит (внутри priority)
-   *
-   * @param tools - Массив инструментов
-   * @returns Отсортированный массив
-   */
-  private sortByPriority(tools: BaseTool[]): BaseTool[] {
-    const priorityOrder: Record<string, number> = {
-      critical: 0,
-      high: 1,
-      normal: 2,
-      low: 3,
-    };
-
-    return tools.sort((a, b) => {
-      // Получаем priority из METADATA
-      const aClass = a.constructor as typeof BaseTool;
-      const bClass = b.constructor as typeof BaseTool;
-      const aPriority = aClass.METADATA?.priority || 'normal';
-      const bPriority = bClass.METADATA?.priority || 'normal';
-
-      const aPrio = priorityOrder[aPriority] ?? 2; // default: normal
-      const bPrio = priorityOrder[bPriority] ?? 2; // default: normal
-
-      // Сначала по priority
-      if (aPrio !== bPrio) {
-        return aPrio - bPrio;
-      }
-
-      // Затем по имени (алфавит)
-      return a.getDefinition().name.localeCompare(b.getDefinition().name);
-    });
-  }
-
-  /**
    * Получить определения всех зарегистрированных инструментов
    *
    * Инструменты отсортированы по приоритету: critical → high → normal → low
@@ -150,28 +114,7 @@ export class ToolRegistry {
     }
 
     const tools = Array.from(this.tools.values());
-    const sorted = this.sortByPriority(tools);
-
-    // Логируем распределение по приоритетам
-    this.logger.debug('Tools sorted by priority', {
-      critical: sorted.filter((t) => {
-        const tClass = t.constructor as typeof BaseTool;
-        return tClass.METADATA?.priority === 'critical';
-      }).length,
-      high: sorted.filter((t) => {
-        const tClass = t.constructor as typeof BaseTool;
-        return tClass.METADATA?.priority === 'high';
-      }).length,
-      normal: sorted.filter((t) => {
-        const tClass = t.constructor as typeof BaseTool;
-        const priority = tClass.METADATA?.priority || 'normal';
-        return priority === 'normal';
-      }).length,
-      low: sorted.filter((t) => {
-        const tClass = t.constructor as typeof BaseTool;
-        return tClass.METADATA?.priority === 'low';
-      }).length,
-    });
+    const sorted = this.sorter.sortByPriority(tools);
 
     return sorted.map((tool) => tool.getDefinition());
   }
@@ -214,7 +157,7 @@ export class ToolRegistry {
       essentialSet.has(tool.getDefinition().name)
     );
 
-    const sorted = this.sortByPriority(tools);
+    const sorted = this.sorter.sortByPriority(tools);
     return sorted.map((tool) => tool.getDefinition());
   }
 
@@ -236,149 +179,11 @@ export class ToolRegistry {
       return this.getDefinitions();
     }
 
-    // Собираем известные категории и подкатегории для валидации
-    const knownCategories = new Set<string>();
-    const knownSubcategories = new Map<string, Set<string>>();
-
-    for (const tool of this.tools.values()) {
-      const toolClass = tool.constructor as typeof BaseTool;
-      const metadata = toolClass.METADATA;
-
-      if (metadata?.category) {
-        knownCategories.add(metadata.category);
-
-        if (metadata.subcategory) {
-          let subcategories = knownSubcategories.get(metadata.category);
-          if (!subcategories) {
-            subcategories = new Set();
-            knownSubcategories.set(metadata.category, subcategories);
-          }
-          subcategories.add(metadata.subcategory);
-        }
-      }
-    }
-
-    // Валидация запрошенных категорий
-    const unknownCategories: string[] = [];
-    const unknownSubcategories: Array<{ category: string; subcategory: string }> = [];
-
-    for (const cat of filter.categories) {
-      if (!knownCategories.has(cat)) {
-        unknownCategories.push(cat);
-      }
-    }
-
-    for (const [cat, subcats] of filter.categoriesWithSubcategories.entries()) {
-      if (!knownCategories.has(cat)) {
-        unknownCategories.push(cat);
-      } else {
-        for (const subcat of subcats) {
-          if (!knownSubcategories.get(cat)?.has(subcat)) {
-            unknownSubcategories.push({ category: cat, subcategory: subcat });
-          }
-        }
-      }
-    }
-
-    // Логируем warnings для неизвестных категорий/подкатегорий
-    if (unknownCategories.length > 0) {
-      this.logger.warn('⚠️  Unknown categories in filter', {
-        unknownCategories: [...new Set(unknownCategories)],
-        knownCategories: Array.from(knownCategories),
-      });
-    }
-
-    if (unknownSubcategories.length > 0) {
-      this.logger.warn('⚠️  Unknown subcategories in filter', {
-        unknownSubcategories,
-        knownSubcategories: Array.from(knownSubcategories.entries()).map(([cat, subcats]) => ({
-          category: cat,
-          subcategories: Array.from(subcats),
-        })),
-      });
-    }
-
-    const tools = Array.from(this.tools.values()).filter((tool) => {
-      const toolClass = tool.constructor as typeof BaseTool;
-      const metadata = toolClass.METADATA;
-
-      if (!metadata?.category) {
-        // Инструменты без категории всегда включены (backwards compatibility)
-        return true;
-      }
-
-      const category = metadata.category;
-      const subcategory = metadata.subcategory;
-
-      // Проверка 1: категория без подкатегории (включает все подкатегории)
-      if (filter.categories.has(category)) {
-        return true;
-      }
-
-      // Проверка 2: категория с конкретными подкатегориями
-      if (subcategory && filter.categoriesWithSubcategories.has(category)) {
-        const allowedSubcategories = filter.categoriesWithSubcategories.get(category);
-        if (allowedSubcategories) {
-          return allowedSubcategories.has(subcategory);
-        }
-      }
-
-      return false;
-    });
-
-    const sorted = this.sortByPriority(tools);
-
-    // Логирование фильтрации
-    this.logger.info('Tools filtered by categories', {
-      totalTools: this.tools.size,
-      filteredTools: sorted.length,
-      categories: Array.from(filter.categories),
-      categoriesWithSubcategories: Array.from(filter.categoriesWithSubcategories.entries()).map(
-        ([cat, subcats]) => ({ category: cat, subcategories: Array.from(subcats) })
-      ),
-    });
+    const allTools = Array.from(this.tools.values());
+    const filtered = this.filterService.filterByCategories(allTools, filter);
+    const sorted = this.sorter.sortByPriority(filtered);
 
     return sorted.map((tool) => tool.getDefinition());
-  }
-
-  /**
-   * Применить негативный фильтр (исключение отключенных групп)
-   *
-   * @param tools - Список инструментов
-   * @param disabledFilter - Фильтр отключенных категорий/подкатегорий
-   * @returns Отфильтрованный список инструментов
-   */
-  private applyDisabledFilter(
-    tools: BaseTool[],
-    disabledFilter: ParsedCategoryFilter
-  ): BaseTool[] {
-    return tools.filter((tool) => {
-      const toolClass = tool.constructor as typeof BaseTool;
-      const metadata = toolClass.METADATA;
-
-      if (!metadata?.category) {
-        // Инструменты без категории всегда включены
-        return true;
-      }
-
-      const category = metadata.category;
-      const subcategory = metadata.subcategory;
-
-      // Проверка 1: категория полностью отключена
-      if (disabledFilter.categories.has(category)) {
-        return false;
-      }
-
-      // Проверка 2: подкатегория отключена
-      if (subcategory && disabledFilter.categoriesWithSubcategories.has(category)) {
-        const disabledSubcategories = disabledFilter.categoriesWithSubcategories.get(category);
-        if (disabledSubcategories?.has(subcategory)) {
-          return false;
-        }
-      }
-
-      return true;
-    });
   }
 
   /**
@@ -403,75 +208,38 @@ export class ToolRegistry {
     }
 
     // Eager mode
+    this.ensureInitialized();
+    if (!this.tools) {
+      return [];
+    }
+
     let tools: BaseTool[];
 
     // Шаг 1: позитивный фильтр (если указан)
     if (categoryFilter && !categoryFilter.includeAll) {
-      this.ensureInitialized();
-      if (!this.tools) {
-        return [];
-      }
-
-      tools = Array.from(this.tools.values()).filter((tool) => {
-        const toolClass = tool.constructor as typeof BaseTool;
-        const metadata = toolClass.METADATA;
-
-        if (!metadata?.category) {
-          return true; // Инструменты без категории всегда включены
-        }
-
-        const category = metadata.category;
-        const subcategory = metadata.subcategory;
-
-        // Проверка категории без подкатегории
-        if (categoryFilter.categories.has(category)) {
-          return true;
-        }
-
-        // Проверка категории с конкретными подкатегориями
-        if (subcategory && categoryFilter.categoriesWithSubcategories.has(category)) {
-          const allowedSubcategories = categoryFilter.categoriesWithSubcategories.get(category);
-          if (allowedSubcategories) {
-            return allowedSubcategories.has(subcategory);
-          }
-        }
-
-        return false;
-      });
+      tools = this.filterService.filterByCategories(
+        Array.from(this.tools.values()),
+        categoryFilter
+      );
     } else {
       // Все инструменты
-      this.ensureInitialized();
-      if (!this.tools) {
-        return [];
-      }
       tools = Array.from(this.tools.values());
     }
 
     // Шаг 2: негативный фильтр (если указан, имеет приоритет)
     if (disabledFilter) {
-      tools = this.applyDisabledFilter(tools, disabledFilter);
-
-      // Логирование отключенных групп
-      this.logger.info('✂️  Применён фильтр отключенных групп', {
-        disabledCategories: Array.from(disabledFilter.categories),
-        disabledCategoriesWithSubcategories: Array.from(
-          disabledFilter.categoriesWithSubcategories.entries()
-        ).map(([cat, subcats]) => ({
-          category: cat,
-          subcategories: Array.from(subcats),
-        })),
-        totalToolsAfterFilter: tools.length,
-      });
+      tools = this.filterService.applyDisabledFilter(tools, disabledFilter);
     }
 
     // Сортировка по приоритету
-    const sorted = this.sortByPriority(tools);
+    const sorted = this.sorter.sortByPriority(tools);
     return sorted.map((tool) => tool.getDefinition());
   }
 
   /**
    * Выполнить инструмент по имени
    */
+  // eslint-disable-next-line max-lines-per-function
   async execute(name: string, params: ToolCallParams): Promise<ToolResult> {
     this.ensureInitialized();
 
@@ -481,6 +249,7 @@ export class ToolRegistry {
     const tool = this.tools?.get(name);
 
     if (!tool) {
+      // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
       const allTools = Array.from(this.tools?.keys() || []);
 
       // Fuzzy поиск похожих имен
